Up: Process class

+++ taskmaster/program.py
@@ -13,6 +13,9 @@ class ProgramProperty:
 		self.required = required
 
 class ProgramParse:
+	'''
+	Parse Yaml configuration
+	'''
 	PARSER={
 		'cmd': ProgramProperty(str, None, True),
 		'numprocs': ProgramProperty(int, 1, False),
@@ -24,6 +27,7 @@ class ProgramParse:
 		'stoptime': ProgramProperty(int, 1, False),
 		'exitcodes': ProgramProperty(list, [0], False),
 		'startretries': ProgramProperty(int, 0, False),
+		'starttime': ProgramProperty(int, -1, False),
 	}
 
 	def parseUni(self, program: dict, prop_name):
@@ -50,6 +54,9 @@ class ProgramParse:
 			self.parseUni(program, prop_name)
 
 class Process(subprocess.Popen):
+	'''
+	Manage Process	
+	'''
 	def __init__(self, name, options):
 		self.name = name
 		self.options = options
@@ -58,24 +65,33 @@ class Process(subprocess.Popen):
 		self.gracefulStop = False
 
 	def myStart(self):
+		if self.start and self.poll() is None:
+			log.Warning(f'{self.name}: Trying to start a process that already running')
+			return f'{self.name}: Trying to start a process that already running\n'
 		null = open('/dev/null', 'r')
 		self.start = True
 		self.gracefulStop = False
-		super().__init__(shlex.split(cmd),
+		super().__init__(shlex.split(self.options.cmd),
 			env=self.options.env,
 			cwd=self.options.workingdir,
 			umask=self.options.umask,
 			stdin=null, stdout=null, stderr=null
 		)
 		self.start_time = datetime.datetime.now()
+		thr = threading.Thread(target=self.myWait, args=(self.options.exitcodes,))
+		thr.start()
+		return f'{self.name}: Started\n'
+
 
 	def myWait(self, exitcodes):
 		returnCode = self.wait()
-		print(returnCode)
-		if returnCode in exitcodes:
-			log.Info(f'{self.name} end successfuly with code {returnCode}')
+		diffTime = datetime.datetime.now() - self.start_time
+		if self.options.starttime >= diffTime.total_seconds():
+			log.Error(f'{self.name}: End badly, fail at {diffTime}')
+		elif returnCode in exitcodes:
+			log.Info(f'{self.name}: End successfuly with code {returnCode}')
 		else:
-			log.Error(f'{self.name} end badly with code {returnCode}')
+			log.Error(f'{self.name}: End badly with code {returnCode}')
 			if not self.gracefulStop and self.retry > 0:
 				log.Info(f'Restart process ({self.name})')
 				self.myStart()
@@ -83,21 +99,26 @@ class Process(subprocess.Popen):
 
 	def myStop(self, stopsignal, stoptime):
 		if not self.start:
-			log.Warning('Trying to stop a process wich not been start')
-			return
+			log.Warning(f'{self.name}: Trying to stop a process wich not been start')
+			return f'{self.name}: Trying to stop a process wich not been start\n'
+		if type(self.poll()) is int:
+			return f'{self.name}: Not running\n'
+		log.Info(f'{self.name}: Graceful Stop')
 		self.send_signal(stopsignal)
 		try:
 			self.wait(stoptime)
 		except subprocess.TimeoutExpired:
 			self.kill()
-			log.Warning('Hard kill. Graceful stop timeout')
+			log.Warning(f'{self.name}: Hard kill. Graceful stop timeout')
+			return f'{self.name}: Hard kill. Graceful stop timeout\n'
+		return f'{self.name}: Stopped\n'
 
 	def get_state(self):
 		return "RUNNING"
 
 	def status(self):
 		if not self.start:
-			return f'{self.name} not start\n'
+			return f'{self.name} not start\n'			
 		uptime = datetime.datetime.now() - self.start_time
 		hours = uptime.total_seconds() // 3600
 		minutes = (uptime.total_seconds() % 3600) // 60
@@ -105,6 +126,9 @@ class Process(subprocess.Popen):
 		return f"{self.name:15}{self.get_state():8} pid {self.pid:6}, uptime {hours:02}:{minutes:02}:{seconds:02}\n"
 
 class Program(ProgramParse):
+	'''
+	Manage list of process
+	'''
 	def __init__(self, program: dict, name: str) -> None:
 		self.parse(program)
 		self.process = []
@@ -125,18 +149,21 @@ class Program(ProgramParse):
 			self.start()
 
 	def start(self):
+		ret = ''
 		for process in self.process:
-			process.myStart()
-			thr = threading.Thread(target=process.myWait, args=(self.exitcodes,))
-			thr.start()
-	
+			ret += process.myStart()
+		return ret
+
 	def stop(self):
+		ret = ''
 		for process in self.process:
-			process.myStop(signals[self.stopsignal], self.stoptime)
+			ret += process.myStop(signals[self.stopsignal], self.stoptime)
+		return ret
 
 	def restart(self):
-		self.stop()
-		self.start()
+		ret = self.stop()
+		ret += self.start()
+		return ret
 	
 	def status(self):
 		status = f"{self.name}: \n"
